[refactor] #15 JACKFLIX ViewController 이름 변경
